+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Switched from CrowdIn to Weblate to get translations for Eco dataset items.
 - Added an ad-hoc recipe calculator for when you want to roughly calculate the
   cost of items that aren't part of your build.
+- Fixed how Barrels are handled so that they're treated as re-integrated
+  products like Molds.
 
 ## 0.2.0 (2026-05-12)